+++ packages/client/components/Tabs/Tabs.tsx
@@ -12,7 +12,7 @@ interface Props {
 const INKBAR_HEIGHT = 2
 
 const InkBar = styled('div')({
-  background: PALETTE.PRIMARY_DARK,
+  background: PALETTE.PRIMARY_MAIN,
   bottom: 0,
   height: INKBAR_HEIGHT,
   left: 0,