+++ packages/client/mutations/NavigateMeetingMutation.ts
@@ -6,7 +6,7 @@ import Atmosphere from '../Atmosphere'
 import {ClientRetrospectiveMeeting} from '../types/clientSchema'
 import {IReflectPhase} from '../types/graphql'
 import {SharedUpdater} from '../types/relayMutations'
-import {VOTE} from '../utils/constants'
+import {REFLECT, VOTE} from '../utils/constants'
 import isInterruptingChickenPhase from '../utils/isInterruptingChickenPhase'
 import isViewerTyping from '../utils/isViewerTyping'
 import getBaseRecord from '../utils/relay/getBaseRecord'
@@ -33,6 +33,9 @@ graphql`
     phaseComplete {
       reflect {
         emptyReflectionGroupIds
+        reflectionGroups {
+          sortOrder
+        }
       }
       group {
         emptyReflectionGroupIds
@@ -100,7 +103,6 @@ export const navigateMeetingTeamUpdater: SharedUpdater<NavigateMeetingMutation_t
     .getValue('id')!
   const meeting = store.get<ClientRetrospectiveMeeting>(meetingId)
   if (!meeting) return
-
   const viewerStageId = safeProxy(meeting)
     .getLinkedRecord('localStage')
     .getValue('id')
@@ -142,6 +144,12 @@ export const navigateMeetingTeamUpdater: SharedUpdater<NavigateMeetingMutation_t
       reflectPrompt?.setValue([], 'editorIds')
     })
   }
+  const reflectionGroups = meeting.getLinkedRecords('reflectionGroups')
+  if (!reflectionGroups) return
+  const sortedReflectionGroups = reflectionGroups
+    .slice()
+    .sort((a, b) => (a.getValue('sortOrder') < b.getValue('sortOrder') ? -1 : 1))
+  meeting.setLinkedRecords(sortedReflectionGroups, 'reflectionGroups')
 }
 
 const NavigateMeetingMutation = (
@@ -175,8 +183,9 @@ const NavigateMeetingMutation = (
         if (stage) {
           stage.setValue(true, 'isComplete')
           const phaseType = stage.getValue('phaseType')
-          if (phaseType === VOTE) {
+          if (phaseType === VOTE || phaseType === REFLECT) {
             // optimistically creating an array of temporary stages is difficult because they can become undefined
+            // same goes for sorting all the reflections based on entities
             // easier to just wait for the return value before advancing
             meeting.setValue(completedStageId, 'facilitatorStageId')
             if (completedStageId) {