+++ packages/server/dataloader/customLoaderMakers.ts
@@ -14,12 +14,14 @@ import MeetingTemplate from '../database/types/MeetingTemplate'
 import {Reactable} from '../database/types/Reactable'
 import Task from '../database/types/Task'
 import {ThreadSource} from '../database/types/ThreadSource'
+import getGitHubAuthByUserIdTeamId, {
+  GetGitHubAuthByUserIdTeamIdResult
+} from '../postgres/queries/getGitHubAuthByUserIdTeamId'
 import AtlassianServerManager from '../utils/AtlassianServerManager'
 import sendToSentry from '../utils/sendToSentry'
 import normalizeRethinkDbResults from './normalizeRethinkDbResults'
 import ProxiedCache from './ProxiedCache'
 import RethinkDataLoader from './RethinkDataLoader'
-
 type TeamUserKey = {teamId: string; userId: string}
 export interface JiraRemoteProjectKey {
   accessToken: string
@@ -213,7 +215,9 @@ export const userTasks = (parent: RethinkDataLoader) => {
               .filter((task) =>
                 archived
                   ? task('tags').contains('archived')
-                  : task('tags').contains('archived').not()
+                  : task('tags')
+                      .contains('archived')
+                      .not()
               )
               .filter((task) => {
                 if (includeUnassigned) return true
@@ -283,6 +287,26 @@ export const freshAtlassianAuth = (parent: RethinkDataLoader) => {
     {
       ...parent.dataLoaderOptions,
       cacheKeyFn: (key) => `${key.userId}:${key.teamId}`
+    }
+  )
+}
+
+export const githubAuth = (parent: RethinkDataLoader) => {
+  return new DataLoader<
+    {teamId: string; userId: string},
+    GetGitHubAuthByUserIdTeamIdResult | null,
+    string
+  >(
+    async (keys) => {
+      const results = await Promise.allSettled(
+        keys.map(async ({teamId, userId}) => getGitHubAuthByUserIdTeamId(userId, teamId))
+      )
+      const vals = results.map((result) => (result.status === 'fulfilled' ? result.value : null))
+      return vals
+    },
+    {
+      ...parent.dataLoaderOptions,
+      cacheKeyFn: ({teamId, userId}) => `${userId}:${teamId}`
     }
   )
 }