+++ packages/server/graphql/mutations/helpers/updateReflectionLocation/removeReflectionFromGroup.ts
@@ -1,4 +1,5 @@
 import getGroupSmartTitle from 'parabol-client/utils/smartGroup/getGroupSmartTitle'
+import dndNoise from '../../../../../client/utils/dndNoise'
 import getRethink from '../../../../database/rethinkDriver'
 import ReflectionGroup from '../../../../database/types/ReflectionGroup'
 import updateSmartGroupTitle from './updateSmartGroupTitle'
@@ -12,16 +13,37 @@ const removeReflectionFromGroup = async (reflectionId, {dataLoader}) => {
     .run()
   if (!reflection) throw new Error('Reflection not found')
   const {reflectionGroupId: oldReflectionGroupId, meetingId, promptId} = reflection
-  const meeting = await dataLoader.get('newMeetings').load(meetingId)
+  const [oldReflectionGroup, reflectionGroupsInColumn, meeting] = await Promise.all([
+    dataLoader.get('retroReflectionGroups').load(oldReflectionGroupId),
+    r
+      .table('RetroReflectionGroup')
+      .getAll(meetingId, {index: 'meetingId'})
+      .filter({isActive: true, promptId})
+      .orderBy('sortOrder')
+      .run(),
+    dataLoader.get('newMeetings').load(meetingId)
+  ])
 
-  // RESOLUTION
-  const reflectionGroup = new ReflectionGroup({meetingId, promptId})
-  await r
-    .table('RetroReflectionGroup')
-    .insert(reflectionGroup)
-    .run()
+  let newSortOrder = 1e6
+  const oldReflectionGroupIdx = reflectionGroupsInColumn.findIndex(
+    (group) => group.id === oldReflectionGroup.id
+  )
+  const sortOrderAtBottom =
+    reflectionGroupsInColumn[reflectionGroupsInColumn.length - 1]?.sortOrder + 1 + dndNoise() ?? 1e6
+  if (oldReflectionGroupIdx === -1 || reflection.promptId !== oldReflectionGroup.promptId) {
+    newSortOrder = sortOrderAtBottom
+  } else if (oldReflectionGroupIdx === reflectionGroupsInColumn.length - 1) {
+    newSortOrder = oldReflectionGroup.sortOrder + 1 + dndNoise()
+  } else {
+    const {sortOrder: oldSortOrder} = oldReflectionGroup
+    const afterSortOrder = reflectionGroupsInColumn[oldReflectionGroupIdx + 1].sortOrder
+    newSortOrder = (oldSortOrder + afterSortOrder) / 2 + dndNoise()
+  }
+
+  const reflectionGroup = new ReflectionGroup({meetingId, promptId, sortOrder: newSortOrder})
   const {id: reflectionGroupId} = reflectionGroup
   await r({
+    reflectionGroup: r.table('RetroReflectionGroup').insert(reflectionGroup),
     reflection: r
       .table('RetroReflection')
       .get(reflectionId)