+++ packages/server/postgres/queries/src/removeGitHubAuthQuery.sql
@@ -1,0 +1,6 @@
+/*
+  @name removeGitHubAuthQuery
+*/
+UPDATE "GitHubAuth"
+SET "isActive" = FALSE, "updatedAt" = CURRENT_TIMESTAMP
+WHERE "userId" = :userId AND "teamId" = :teamId AND "isActive" = TRUE;