+++ packages/server/postgres/queries/upsertGitHubAuth.ts
@@ -1,0 +1,10 @@
+import getPg from '../getPg'
+import {
+  IUpsertGitHubAuthQueryParams,
+  upsertGitHubAuthQuery
+} from './generated/upsertGitHubAuthQuery'
+
+const upsertGitHubAuth = async (auth: IUpsertGitHubAuthQueryParams['auth']) => {
+  await upsertGitHubAuthQuery.run({auth}, getPg())
+}
+export default upsertGitHubAuth